+++ src/components/common/index.ts
@@ -1,2 +1,4 @@
 export { default as Button } from './Button';
 export { default as PageTitle } from './PageTitle';
+export { default as SignInput } from './SignInput';
+export { default as SignContainer } from './SignContainer';
